requirements.txt
torch>=2.2
torchvision
torchaudio

transformers>=4.37.0
datasets>=2.18.0
tokenizers>=0.15.0
accelerate>=0.28.0
deepspeed
peft>=0.10.0
trl>=0.16.0

sentencepiece
safetensors
scikit-learn
numpy
pandas
pyarrow
ujson
jsonlines
tqdm
pyyaml
tensorboard
evaluate
jieba
nltk
